+++ migrations/2021-09-21-081512_skip_checks/up.sql
@@ -1,0 +1,2 @@
+ALTER TABLE repository ADD COLUMN default_enable_qa BOOL NOT NULL default true;
+ALTER TABLE repository ADD COLUMN default_enable_checks BOOL NOT NULL default true;